Improved provisioning allowing it to be run multiple times

+++ scripts/bootstrap.sh
@@ -14,7 +14,9 @@ $IMUxSockRateLimitSeverity 7
 EOF
 service rsyslog restart
 }
-[ `grep 'mirrors.kernel.org' /etc/yum.repos.d/CentOS-Base.repo` ] || {
-    unalias cp 2>/dev/null || /bin/true
-    cp -f /vagrant/scripts/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo
-}
+if [ -e /etc/yum.repos.d/CentOS-Base.repo ]; then
+    [ `grep 'mirrors.kernel.org' /etc/yum.repos.d/CentOS-Base.repo` ] || {
+        unalias cp 2>/dev/null || /bin/true
+        cp -f /vagrant/scripts/CentOS-Base.repo /etc/yum.repos.d/CentOS-Base.repo
+    }
+fi
